# api-gateway/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AeroLink Distributed Platform API",
    "version": "2.0.0",
    "description": "AeroLink Airline Systems Platform – Multi-Region Cloud API. Implements RESTful endpoints across 4 microservices: Booking (Aurora SQL), Baggage (DynamoDB NoSQL), Check-In, and Flight Operations. Secured with JWT/OAuth2 and RBAC.",
    "contact": { "name": "AeroLink Engineering", "email": "api@aerolink.io" },
    "license": { "name": "MIT" }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local Development Gateway" },
    { "url": "https://api.aerolink.io", "description": "AWS Production (US-East-1)" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "Flight": {
        "type": "object",
        "properties": {
          "flight_id": { "type": "string", "example": "AL-101" },
          "origin": { "type": "string", "example": "Colombo (CMB)" },
          "destination": { "type": "string", "example": "London (LHR)" },
          "available_seats": { "type": "integer", "example": 145 },
          "total_seats": { "type": "integer", "example": 150 },
          "departure_time": { "type": "string", "format": "date-time" },
          "arrival_time": { "type": "string", "format": "date-time" },
          "price": { "type": "number", "example": 450.00 },
          "status": { "type": "string", "enum": ["ON_TIME", "DELAYED", "BOARDING", "CANCELLED"] }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/bookings/flights": {
      "get": {
        "summary": "List all available flights",
        "description": "Retrieves relational flight data from Aurora SQL simulation.",
        "tags": ["Booking Service"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Array of flights",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Flight" } } } } } }
          }
        }
      }
    },
    "/api/bookings/reserve": {
      "post": {
        "summary": "Reserve a flight seat (ACID transaction)",
        "description": "Atomically decrements seat count and creates booking record. Emits SEAT_RESERVED event on SSE stream.",
        "tags": ["Booking Service"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["flightId", "passengerName"],
                "properties": {
                  "flightId": { "type": "string", "example": "AL-101" },
                  "passengerName": { "type": "string", "example": "James Smith" },
                  "seatClass": { "type": "string", "enum": ["Economy", "Business", "First"], "example": "Economy" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Booking confirmed with assigned seat number" },
          "400": { "description": "No seats available or missing fields" },
          "404": { "description": "Flight not found" }
        }
      }
    },
    "/api/bookings/all": {
      "get": {
        "summary": "List all bookings (Staff/Admin)",
        "tags": ["Booking Service"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "All bookings joined with flight data" } }
      }
    },
    "/api/bookings/{id}": {
      "get": {
        "summary": "Get booking by ID",
        "tags": ["Booking Service"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "example": "BK-54321" }],
        "responses": {
          "200": { "description": "Booking details" },
          "404": { "description": "Booking not found" }
        }
      }
    },
...
  }
}